+++ Cockpybara/src/main/java/Alchole_free/Cockpybara/domain/member/Gender.java
@@ -1,4 +1,4 @@
-package Alchole_free.Cockpybara.domain;
+package Alchole_free.Cockpybara.domain.member;
 
 public enum Gender {
     MALE, FEMALE
